Split Lua symbol tests for #3383
